requirements.txt
click==7.1.1
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pytube3==9.6.4
typing-extensions==3.7.4.2
wget==3.2
